# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Media API",
    "version": "1.0.0",
    "description": "API for Social Media Application"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "user posts",
        "description": "a list of posts by currrent user",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numerical value for page"
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "add post",
        "description": "adds a new post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created"
          },
          "404": {
            "description": "post not found"
          }
        }
      }
    },
...
  }
}